Some more rearranging.

+++ RasterPropMonitor/UtilityFunctions.cs
@@ -89,7 +89,7 @@ namespace JSI
 			return SignedAngle(Vector3.Cross(up, a), Vector3.Cross(up, b), up);
 		}
 
-		public static float SignedAngle(Vector3 v1, Vector3 v2, Vector3 up)
+		public static double SignedAngle(Vector3 v1, Vector3 v2, Vector3 up)
 		{
 			if (Vector3.Dot(Vector3.Cross(v1, v2), up) < 0)
 				return -Vector3.Angle(v1, v2);
